feat: implement Navbar component with active link handling and nested links
